refactor(server): split cross review tests
